feat(ProviderContainer): Add navigation for Invest and Manage pools
