+++ app/package.json
@@ -24,7 +24,7 @@
     "web3": "^1.3.0"
   },
   "scripts": {
-    "start": "react-app-rewired start",
+    "start": "HTTPS=true SSL_CRT_FILE=./.cert/cert.pem SSL_KEY_FILE=./.cert/key.pem react-app-rewired start",
     "build": "react-app-rewired build",
     "test": "react-app-rewired test",
     "eject": "react-scripts eject",
