--- app/src/components/index.ts
@@ -1,7 +1,0 @@
-import SelectedFund from './SelectedFund'
-import SelectFund from './SelectFund'
-
-export {
-    SelectFund,
-    SelectedFund
-}
