+++ app/src/components/investor/Invest.tsx
