+++ app/src/components/investor/SelectedFund.tsx
