+++ app/src/components/investor/index.ts
@@ -1,0 +1,11 @@
+import SelectedFund from './SelectedFund';
+import SelectFund from './SelectFund';
+import Invest from './Invest';
+import Withdraw from './Withdraw';
+
+export {
+    Invest,
+    Withdraw,
+    SelectFund,
+    SelectedFund
+}
