+++ app/src/containers/index.ts
@@ -1,0 +1,7 @@
+import InvestorContainer from './InvestorContainer';
+import ProviderContainer from "./ProviderContainer";
+
+export {
+  InvestorContainer,
+  ProviderContainer,
+}
